See DC42 1.09h commit for more details

https://github.com/dc42/RepRapFirmware/commit/13c1fc77fb81bd16b265e53e6108e2e344f5c9f4

This version retains support for all the Duet v0.8.5 changes to support
2 fans and 6 extruders but otherwise is identical to DC42 1.09h.

+++ GCodeBuffer.cpp
@@ -114,7 +114,7 @@ bool GCodeBuffer::Put(char c)
 	else if (!inComment || writingFileDirectory)
 	{
 		gcodeBuffer[gcodePointer++] = c;
-		if (gcodePointer >= GcodeLength)
+		if (gcodePointer >= (int)GcodeLength)
 		{
 			platform->Message(BOTH_ERROR_MESSAGE, "G-Code buffer length overflow.\n");
 			gcodePointer = 0;
